requirements.txt
# ========================================
# Core Dependencies
# ========================================
python-dotenv==1.0.0
pydantic==2.5.0

# ========================================
# FastAPI & Web Server
# ========================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1

# ========================================
# Machine Learning
# ========================================
tensorflow==2.15.0
numpy==1.24.3
scikit-learn==1.3.2
pillow==10.1.0

# ========================================
# Data Processing
# ========================================
pandas==2.1.3
rasterio==1.3.9
scikit-image==0.22.0

# ========================================
# Drift Detection
# ========================================
evidently==0.5.0

# ========================================
# Google Cloud & Earth Engine
# ========================================
earthengine-api==0.1.382
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
google-cloud-storage==2.13.0

# ========================================
# Utilities
# ========================================
requests==2.32.0
python-dateutil==2.8.2
pyyaml==6.0.1
tqdm==4.66.1

# ========================================
# Testing
# ========================================
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# ========================================
# Development Tools
# ========================================
black==23.11.0
flake8==6.1.0
mypy==1.7.1
